+++ .gitignore
@@ -12,6 +12,7 @@
 /.idea/gradle.xml
 /.idea/misc.xml
 /.idea/jarRepositories.xml
+/.idea/codeStyles
 .idea/workspace.xml
 .idea/tasks.xml
 .idea/gradle.xml
