+++ app/src/main/java/com/yml/womensafety/repository/FirebaseRepository.kt
@@ -19,7 +19,6 @@ class FirebaseRepository {
     private val appUser = FirebaseUtil.user?.currentUser
     private val userReference = databaseReference?.child(appUser?.uid!!)
     fun getUserName(response: Response) {
-        FirebaseUtil.firebaseDatabase?.setPersistenceEnabled(true)
         databaseReference?.keepSynced(true)
         userReference?.addValueEventListener(object : ValueEventListener {
             override fun onDataChange(snapshot: DataSnapshot) {
