+++ app/src/main/res/layout/fragment_contacts.xml
@@ -38,7 +38,6 @@
             android:layout_width="@dimen/contacts_layout_width_height"
             android:layout_height="wrap_content"
             android:layout_marginTop="@dimen/contact_name_margin_top"
-            android:background="@drawable/edit_text_shape"
             android:hint="@string/name"
             android:inputType="text"
             android:padding="@dimen/et_padding"
@@ -54,7 +53,6 @@
             android:layout_width="@dimen/contacts_layout_width_height"
             android:layout_height="wrap_content"
             android:layout_marginTop="@dimen/contact_number_margin_top"
-            android:background="@drawable/edit_text_shape"
             android:hint="@string/phone_number"
             android:inputType="number"
             android:maxLength="10"
