+++ app/src/main/res/layout/fragment_introduction_contacts.xml
@@ -18,7 +18,7 @@
         app:layout_constraintVertical_bias=".3" />
 
     <TextView
-        android:id="@+id/textView"
+        android:id="@+id/tvTitleLabel"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="@dimen/ss_text_view_margin_top"
@@ -42,14 +42,14 @@
         android:textColor="@color/black"
         android:textSize="@dimen/ss_title_text_size"
         android:textStyle="bold"
-        app:layout_constraintBottom_toBottomOf="@+id/textView"
+        app:layout_constraintBottom_toBottomOf="@+id/tvTitleLabel"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="1"
-        app:layout_constraintStart_toEndOf="@+id/textView"
-        app:layout_constraintTop_toTopOf="@+id/textView" />
+        app:layout_constraintStart_toEndOf="@+id/tvTitleLabel"
+        app:layout_constraintTop_toTopOf="@+id/tvTitleLabel" />
 
     <TextView
-        android:id="@+id/textView2"
+        android:id="@+id/tvTitleTwo"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:fontFamily="sans-serif-light"
@@ -64,7 +64,7 @@
         app:layout_constraintVertical_bias="0" />
 
     <TextView
-        android:id="@+id/textView3"
+        android:id="@+id/tvTitleThree"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:fontFamily="sans-serif-condensed"
@@ -75,12 +75,12 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0"
-        app:layout_constraintStart_toStartOf="@id/textView2"
-        app:layout_constraintTop_toBottomOf="@id/textView2"
+        app:layout_constraintStart_toStartOf="@id/tvTitleTwo"
+        app:layout_constraintTop_toBottomOf="@id/tvTitleTwo"
         app:layout_constraintVertical_bias="0" />
 
     <TextView
-        android:id="@+id/textView4"
+        android:id="@+id/tvDescription"
         android:layout_width="@dimen/ss_text_description_width_height"
         android:layout_height="@dimen/ss_text_description_width_height"
         android:layout_marginTop="@dimen/ss_text_view_4_margin_top"
@@ -91,8 +91,8 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHeight_percent=".1"
         app:layout_constraintHorizontal_bias="0"
-        app:layout_constraintStart_toStartOf="@+id/textView3"
-        app:layout_constraintTop_toBottomOf="@id/textView3"
+        app:layout_constraintStart_toStartOf="@+id/tvTitleThree"
+        app:layout_constraintTop_toBottomOf="@id/tvTitleThree"
         app:layout_constraintVertical_bias="0"
         app:layout_constraintWidth_percent=".5" />
 
@@ -104,8 +104,8 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0"
-        app:layout_constraintStart_toStartOf="@+id/textView4"
-        app:layout_constraintTop_toBottomOf="@+id/textView4"
+        app:layout_constraintStart_toStartOf="@+id/tvDescription"
+        app:layout_constraintTop_toBottomOf="@+id/tvDescription"
         app:layout_constraintVertical_bias=".6" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
